stopping here
code works; but there's a problem.
EEE-HPC doesn't match the reports. The reports doesn't match the reports. The geometry doesn't match the reports. All of this was done before I was born so I have no idea what is going on or who has the data for EEE. People who made EEE the few that are left - why wasn't anything documented properly.